+++ jumanji/environments/packing/job_shop/constructive/env.py
@@ -23,8 +23,8 @@ from numpy.typing import NDArray
 
 from jumanji import specs
 from jumanji.env import Environment
-from jumanji.environments.packing.job_shop.generator import Generator, RandomGenerator
-from jumanji.environments.packing.job_shop.types import Observation, State
+from jumanji.environments.packing.job_shop.constructive.generator import Generator, RandomGenerator
+from jumanji.environments.packing.job_shop.constructive.types import Observation, State
 from jumanji.environments.packing.job_shop.viewer import JobShopViewer
 from jumanji.types import TimeStep, restart, termination, transition
 from jumanji.viewer import Viewer
