+++ jumanji/environments/packing/job_shop/types.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import TYPE_CHECKING, NamedTuple, Optional
+from typing import TYPE_CHECKING
 
 import chex
 
@@ -22,60 +22,50 @@ else:
     from chex import dataclass
 
 
-class Observation(NamedTuple):
-    """
+@dataclass
+class Scenario:
+    """A scenario containing the description of the job shop scheduling problem.
+    It is common to the constructive and improvement methods.
+
+    num_jobs: number of jobs in the problem.
+    num_machines: number of machines in the problem.
+    max_num_jobs: maximum number of jobs in the problem.
+    max_num_ops: maximum number of operations in the problem.
     ops_machine_ids: for each job, it specifies the machine each op must be processed on.
         Note that a -1 corresponds to padded ops since not all jobs have the same number of ops.
     ops_durations: for each job, it specifies the processing time of each operation.
         Note that a -1 corresponds to padded ops since not all jobs have the same number of ops.
-    ops_mask: for each job, indicates which operations remain to be scheduled. False if the
-        op has been scheduled or if the op was added for padding, True otherwise. The first True in
-        each row (i.e. each job) identifies the next operation for that job.
-    machines_job_ids: for each machine, it specifies the job currently being processed. Note that
-        the index num_jobs represents a no-op for which the time until available is always 0.
-    machines_remaining_times: for each machine, it specifies the number of time steps until
-        available.
-    action_mask: for each machine, it indicates which jobs (or no-op) can legally be scheduled.
-        The last column corresponds to no-op.
+    num_ops_per_job: for each job, it specifies the number of operations.
+        Note that a 0 corresponds to a non-existing job.
     """
 
-    ops_machine_ids: chex.Array  # (num_jobs, max_num_ops)
-    ops_durations: chex.Array  # (num_jobs, max_num_ops)
-    ops_mask: chex.Array  # (num_jobs, max_num_ops)
-    machines_job_ids: chex.Array  # (num_machines,)
-    machines_remaining_times: chex.Array  # (num_machines,)
-    action_mask: chex.Array  # (num_machines, num_jobs + 1)
+    num_jobs: int
+    num_machines: int
+    max_num_jobs: int
+    max_num_ops: int
+    ops_machine_ids: chex.Array  # (max_num_jobs, max_num_ops)
+    ops_durations: chex.Array  # (max_num_jobs, max_num_ops)
+    num_ops_per_job: chex.Array  # (max_num_jobs,)
 
 
 @dataclass
-class State:
+class CommonState:
     """The environment state containing a complete description of the job shop scheduling problem.
+    It is common to the constructive and improvement methods.
+    States of each method are derived from this class.
 
     ops_machine_ids: for each job, it specifies the machine each op must be processed on.
         Note that a -1 corresponds to padded ops since not all jobs have the same number of ops.
     ops_durations: for each job, it specifies the processing time of each operation.
         Note that a -1 corresponds to padded ops since not all jobs have the same number of ops.
-    ops_mask: for each job, indicates which operations remain to be scheduled. False if the
-        op has been scheduled or if the op was added for padding, True otherwise. The first True in
-        each row (i.e. each job) identifies the next operation for that job.
-    machines_job_ids: for each machine, it specifies the job currently being processed. Note that
-        the index num_jobs represents a no-op for which the time until available is always 0.
-    machines_remaining_times: for each machine, it specifies the number of time steps until
-        available.
-    action_mask: for each machine, it indicates which jobs (or no-op) can legally be scheduled.
-        The last column corresponds to no-op.
     step_count: used to track time, which is necessary for updating scheduled_times.
     scheduled_times: for each job, it specifies the time at which each operation was scheduled.
         Note that -1 means the operation has not been scheduled yet.
     key: random key used for auto-reset.
     """
 
-    ops_machine_ids: chex.Array  # (num_jobs, max_num_ops)
-    ops_durations: chex.Array  # (num_jobs, max_num_ops)
-    ops_mask: chex.Array  # (num_jobs, max_num_ops)
-    machines_job_ids: chex.Array  # (num_machines,)
-    machines_remaining_times: chex.Array  # (num_machines,)
-    action_mask: Optional[chex.Array]  # (num_machines, num_jobs + 1)
+    ops_machine_ids: chex.Array  # (max_num_jobs, max_num_ops)
+    ops_durations: chex.Array  # (max_num_jobs, max_num_ops)
     step_count: chex.Numeric  # ()
-    scheduled_times: chex.Array  # (num_jobs, max_num_ops)
+    scheduled_times: chex.Array  # (max_num_jobs, max_num_ops)
     key: chex.PRNGKey  # (2,)
