+++ jumanji/training/networks/job_shop/actor_critic.py
@@ -20,8 +20,8 @@ import jax
 import jax.numpy as jnp
 import numpy as np
 
-from jumanji.environments.packing.job_shop import JobShop
-from jumanji.environments.packing.job_shop.types import Observation
+from jumanji.environments.packing.job_shop.constructive.env import JobShop
+from jumanji.environments.packing.job_shop.constructive.types import Observation
 from jumanji.training.networks.actor_critic import (
     ActorCriticNetworks,
     FeedForwardNetwork,
